Recover CastSpellFromSource without gotos; type the effect's original target
Restore retail's fallback "Default" camera visual and GeneralMagic.wav for
unknown target types, use TransformTransposed for aimed effects as retail
0x004fb4c0 does, and name the spell effect's OrigTarget slot that the save
chunk already stores after OrigSource.

+++ src/wiz8/local_code/PC Item.cpp
@@ -4050,7 +4050,7 @@ int CastItemSpell0051EE70(W8Character* character, W8ItemInstance* item, unsigned
         int power_out = static_cast<int>(power);
         effect = CastSpellFromSource(
             spell_id, &target, &g_status_685170.buffers.XChar[party_slot].target_out_of_combat,
-            power, 0, static_cast<int>(difficulty), 0, &power_out, difficulty_kind, 0, 0);
+            power, 0, difficulty, 0, &power_out, difficulty_kind, 0, 0);
         power = static_cast<unsigned int>(power_out);
     }
